refactor(server): extract session tools into src/server/handlers/session-tools.ts
create_debug_session, list_debug_sessions and close_debug_session move off
the CallTool switch as ToolHandler functions taking the live ToolContext;
bodies verbatim with `this.` -> `ctx.` and `result = X; break;` ->
`return X;`. The three branch-terminal assignments in create_debug_session's
attach/launch arms (try / catch / else, each the last statement before the
case's trailing break) become returns for the same reason. The switch cases
now delegate to the new functions until tool-dispatch replaces the switch.

handleListDebugSessions moves too; the class keeps a one-line private
delegate as the test seam (removed in PR 6).
